renamed and repackaged home repo
