+++ daggercore/src/main/java/com/worldsnas/daggercore/modules/RepoModule.kt
@@ -1,13 +1,11 @@
 package com.worldsnas.daggercore.modules
 
-import com.worldsnas.domain.repo.discover.DiscoverModule
-import com.worldsnas.domain.repo.featured.FeaturedModule
+import com.worldsnas.domain.repo.home.HomeModule
 import dagger.Module
 
 @Module(
     includes = [
-        DiscoverModule::class,
-        FeaturedModule::class
+        HomeModule::class
     ]
 )
 abstract class RepoModule
