--- domain/src/main/java/com/worldsnas/domain/repo/discover/DiscoverModule.kt
@@ -1,8 +1,0 @@
-package com.worldsnas.domain.repo.discover
-
-import dagger.Module
-
-@Module
-abstract class DiscoverModule {
-
-}
