--- domain/src/main/java/com/worldsnas/domain/repo/discover/DiscoverRepo.kt
@@ -1,4 +1,0 @@
-package com.worldsnas.domain.repo.discover
-
-interface DiscoverRepo {
-}
