--- domain/src/main/java/com/worldsnas/domain/repo/featured/FeaturedAPI.kt
@@ -1,4 +1,0 @@
-package com.worldsnas.domain.repo.featured
-
-interface FeaturedAPI {
-}
