--- domain/src/main/java/com/worldsnas/domain/repo/featured/FeaturedModule.kt
@@ -1,8 +1,0 @@
-package com.worldsnas.domain.repo.featured
-
-import dagger.Module
-
-@Module
-abstract class FeaturedModule {
-
-}
