--- domain/src/main/java/com/worldsnas/domain/repo/featured/FeaturedRepo.kt
@@ -1,4 +1,0 @@
-package com.worldsnas.domain.repo.featured
-
-interface FeaturedRepo {
-}
