--- domain/src/main/java/com/worldsnas/domain/repo/featured/FeaturedRepoImpl.kt
@@ -1,6 +1,0 @@
-package com.worldsnas.domain.repo.featured
-
-import javax.inject.Inject
-
-class FeaturedRepoImpl @Inject constructor(
-): FeaturedRepo
