+++ domain/src/main/java/com/worldsnas/domain/repo/home/HomeModule.kt
@@ -1,0 +1,8 @@
+package com.worldsnas.domain.repo.home
+
+import dagger.Module
+
+@Module
+abstract class HomeModule {
+
+}
