+++ domain/src/main/java/com/worldsnas/domain/repo/home/discover/LatestMovieRepo.kt
@@ -1,0 +1,4 @@
+package com.worldsnas.domain.repo.home.discover
+
+interface LatestMovieRepo {
+}
