+++ domain/src/main/java/com/worldsnas/domain/repo/home/trending/TrendingFetcher.kt
@@ -1,10 +1,10 @@
-package com.worldsnas.domain.repo.featured
+package com.worldsnas.domain.repo.home.trending
 
 import com.worldsnas.panther.RFetcher
 import retrofit2.Response
 import javax.inject.Inject
 
-class FeaturedFetcher @Inject constructor(
+class TrendingFetcher @Inject constructor(
 ) : RFetcher<String, Int>{
     override fun fetch(param: String): Response<Int> {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
