+++ domain/src/main/java/com/worldsnas/domain/repo/home/trending/TrendingRepo.kt
@@ -1,0 +1,4 @@
+package com.worldsnas.domain.repo.home.trending
+
+interface TrendingRepo {
+}
