+++ domain/src/main/java/com/worldsnas/domain/repo/home/trending/TrendingRepoImpl.kt
@@ -1,0 +1,6 @@
+package com.worldsnas.domain.repo.home.trending
+
+import javax.inject.Inject
+
+class TrendingRepoImpl @Inject constructor(
+): TrendingRepo
